# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend Go + Gin untuk aplikasi video conference berbasis LiveKit.\nAuth pakai dua jenis JWT: app JWT (untuk endpoint backend ini) dan LiveKit JWT (untuk connect ke LiveKit server, di-issue oleh /api/token).",
        "title": "Videoconf Backend API",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/recordings/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Owner only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recordings"
                ],
                "summary": "Detail recording",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "recording id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/videoconf-backend_internal_models.Recording"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/recordings/{id}/stop": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Owner only (cek via room.owner_id). Backend call Egress stop, update status ke \"ending\". Status jadi \"complete\" pas file selesai upload (sekarang manual update — webhook auto-update bisa di-add nanti).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recordings"
                ],
                "summary": "Stop recording",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "recording id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/videoconf-backend_internal_models.Recording"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/rooms/{idOrSlug}/recordings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Owner only. Urut paling baru duluan (DESC by started_at).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recordings"
                ],
                "summary": "List recording per room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "room id atau slug",
                        "name": "idOrSlug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.recordingsListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Owner only. Pakai LiveKit Egress room composite (grid layout) → MP4 → upload ke MinIO bucket \"recordings\". Egress butuh participant aktif — kalau room kosong return 502.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recordings"
                ],
                "summary": "Mulai recording room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "room id atau slug",
                        "name": "idOrSlug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "status: starting",
                        "schema": {
                            "$ref": "#/definitions/videoconf-backend_internal_models.Recording"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "403": {
                        "description": "bukan owner",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "502": {
                        "description": "Egress error (room kosong, dll)",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handlers.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "human-readable error message"
                }
            }
        },
...
        "internal_handlers.recordingsListResponse": {
            "type": "object",
            "properties": {
                "recordings": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "duration_seconds": {
                                "type": "integer"
                            },
                            "egress_id": {
                                "type": "string"
                            },
                            "ended_at": {
                                "type": "string"
                            },
                            "error": {
                                "type": "string"
                            },
                            "file_path": {
                                "type": "string"
                            },
                            "file_size": {
                                "type": "integer"
                            },
                            "file_url": {
                                "type": "string"
                            },
                            "id": {
                                "type": "integer"
                            },
                            "room_id": {
                                "type": "integer"
                            },
                            "started_at": {
                                "type": "string"
                            },
                            "started_by": {
                                "type": "integer"
                            },
                            "status": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "videoconf-backend_internal_models.Recording": {
            "type": "object",
            "properties": {
                "duration_seconds": {
                    "type": "integer"
                },
                "egress_id": {
                    "type": "string"
                },
                "ended_at": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "file_path": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "file_url": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "room_id": {
                    "type": "integer"
                },
                "started_at": {
                    "type": "string"
                },
                "started_by": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" lalu spasi, lalu app JWT yang didapat dari /auth/login atau /auth/register.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}